pyproject.toml
[project]
name = "bsl-excel"
version = "0.1.0"
description = "Query a DuckDB semantic layer from Excel via FastAPI"
requires-python = ">=3.11"
dependencies = [
    "boring-semantic-layer>=0.3.2",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "ibis-framework[duckdb]>=9.0",
    "duckdb>=1.4.4"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
